@@ -12,18 +12,19 @@
     "@ai-sdk/openai": "^1.0.11",
     "@date-fns/tz": "^1.2.0",
     "@hookform/resolvers": "^3.9.1",
-    "@novu/headless": "^2.0.3",
+    "@novu/headless": "^2.0.4",
     "@tanstack/react-table": "^8.20.6",
     "@upstash/ratelimit": "^2.0.5",
     "ai": "^4.0.22",
     "dub": "^0.46.8",
     "framer-motion": "^11.15.0",
     "geist": "^1.3.1",
-    "next": "^15.1.1-canary.16",
+    "next": "^15.1.1-canary.21",
+    "next-auth": "^5.0.0-beta.25",
     "next-international": "^1.3.1",
     "next-safe-action": "^7.10.2",
     "next-themes": "^0.4.4",
-    "novu": "^2.2.1",
+    "novu": "^2.2.2",
     "nuqs": "^2.2.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
@@ -33,17 +34,17 @@
     "react-use-draggable-scroll": "^0.4.7",
     "resend": "^4.0.1",
     "sonner": "^1.7.1",
+    "stripe": "^17.5.0",
     "use-long-press": "^3.2.0",
     "zustand": "^5.0.2"
   },
   "devDependencies": {
-    "typescript": "^5",
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "postcss": "^8",
-    "tailwindcss": "^3.4.1",
-    "@bubba-beta/db": "workspace:*",
-    "@bubba-beta/auth": "workspace:*"
+    "typescript": "^5.7.2",
+    "@types/node": "^22.10.2",
+    "@types/react": "^19.0.2",
+    "@types/react-dom": "^19.0.2",
+    "postcss": "^8.4.49",
+    "tailwindcss": "^3.4.17",
+    "@bubba-beta/db": "workspace:*"
   }
 }
