@@ -14,7 +14,7 @@
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
     "turbo": "^2.3.3",
-    "typescript": "5.5.4"
+    "typescript": "5.7.2"
   },
   "engines": {
     "node": ">=18"
