@@ -1,8 +1,8 @@
+import { auth } from "@/auth";
 import { DeleteOrganization } from "@/components/forms/organization/delete-organization";
 import { UpdateOrganizationName } from "@/components/forms/organization/update-organization-name";
 import { UpdateOrganizationWebsite } from "@/components/forms/organization/update-organization-website";
 import { SeedDataForm } from "@/components/forms/seed-data-form";
-import { auth } from "@bubba-beta/auth";
 import { db } from "@bubba-beta/db";
 import type { Metadata } from "next";
 import { redirect } from "next/navigation";
