@@ -4,7 +4,7 @@ export function useEnterSubmit(): {
   formRef: RefObject<HTMLFormElement>;
   onKeyDown: (event: React.KeyboardEvent<HTMLTextAreaElement>) => void;
 } {
-  const formRef = useRef<HTMLFormElement>(null);
+  const formRef = useRef<HTMLFormElement>(null) as RefObject<HTMLFormElement>;
 
   const handleKeyDown = (
     event: React.KeyboardEvent<HTMLTextAreaElement>,
