@@ -400,7 +400,7 @@ export const Icons = {
       />
     </svg>
   ),
-  FileDocument: (props) => (
+  FileDocument: (props: any) => (
     <svg
       xmlns="http://www.w3.org/2000/svg"
       width={24}
