@@ -1,18 +1,16 @@
-import { type RefObject, useRef } from 'react';
+import { type RefObject, useRef } from "react";
 
 export function useEnterSubmit(): {
   formRef: RefObject<HTMLFormElement>;
-  onKeyDown: (
-    event: React.KeyboardEvent<HTMLTextAreaElement>
-  ) => void;
+  onKeyDown: (event: React.KeyboardEvent<HTMLTextAreaElement>) => void;
 } {
-  const formRef = useRef<HTMLFormElement>(null);
+  const formRef = useRef<HTMLFormElement>(null) as RefObject<HTMLFormElement>;
 
   const handleKeyDown = (
-    event: React.KeyboardEvent<HTMLTextAreaElement>
+    event: React.KeyboardEvent<HTMLTextAreaElement>,
   ): void => {
     if (
-      event.key === 'Enter' &&
+      event.key === "Enter" &&
       !event.shiftKey &&
       !event.nativeEvent.isComposing
     ) {
